+++ POOTrabalhoPrat/POOTrabalhoPrat/Alianca_Diplomatica.h
@@ -4,9 +4,8 @@
 
 class Alianca_Diplomatica : public Evento {
 private:
-
 public:
 	Alianca_Diplomatica();
 	~Alianca_Diplomatica();
-	void efeito(Imperio_Jogador *const, const int) override;
+	void efeito(Imperio_Jogador *const, const float) override;
 };
