+++ POOTrabalhoPrat/POOTrabalhoPrat/Invasao.cpp
@@ -1,4 +1,5 @@
 #include "Invasao.h"
+#include "Imperio_Jogador.h"
 
 using namespace std;
 
@@ -6,10 +7,10 @@ Invasao::Invasao() : Evento("Invasao", "Um outro imperio concorrente esta a tent
 
 Invasao::~Invasao() {}
 
-void Invasao::efeito(Imperio_Jogador *const imp, const int ano) {
+void Invasao::efeito(Imperio_Jogador *const imp, const float ano) {
 	Evento::efeito(imp, ano);
 	int fator_sorte = imp->gerar_fator_sorte();
-	fator_sorte += ano == 2 ? 2 : 3;
+	fator_sorte += ano == 2.0f ? 2 : 3;
 	bool invadido = imp->territorio_invadido(fator_sorte);
 	cout << (invadido ? "Territorio invadido" : "Territorio nao invadido") << endl;
 }
