+++ POOTrabalhoPrat/POOTrabalhoPrat/Recurso_Abandonado.h
@@ -4,9 +4,8 @@
 
 class Recurso_Abandonado : public Evento {
 private:
-
 public:
 	Recurso_Abandonado();
 	~Recurso_Abandonado();
-	void efeito(Imperio_Jogador *const, const int) override;
+	void efeito(Imperio_Jogador *const, const float) override;
 };
